Try to track down reason of VTE-related crash.
As it seems, VTE terminals (xfce4-terminal, roxterm, etc.) create 4
windows upon starting, but only map one of them to the root window.

This is not a problem per se, however upon closing such a terminal, only
3 of the 4 are destroyed (and one of the windows is destroyed twice).
This causes problems in scan(), since there is a window left that is
probably not supposed to be there.

=> need to find out why VTE does this, and what dwm does to solve this
   issue.