Fix error message upon success.
The error message was displayed at the end of a loop (when the loop does
not return upon finding the right window).
Problem is there was a 'break' instead of a 'return', so the error
message got displayed even upon successful deletion of the correct
client.

Move around config variables (might get moved to a separate config.h
file later), and arranged function declarations alphabetically.