Set MODKEY to Mod1 (Alt), customise xinitrc, update README.
xinitrc now calls dispwin to calibrate the screen, which doesn't work in
Xephyr (one of the reasons why I'm developing now on a real X display;
see the last commit).

There is no longer a need for keeping the default MODKEY to the Windows-
key for avoid conflicts with my (current) main window manager; so
switching to the more confortable Alt-key.