Draw a status bar at the top, start handling fonts.
The status bar doesn't yet claim an area on the screen for itself, and
it doesn't yet show anything useful (just a test string).

Given that with a simple XDrawString(), font ascents/descents are not
well handled, positioning a text is quite cumbersome, hence I'm starting
with some XFont stuff now.