Add workspace dimensions and font handling.
Workspace dimensions are now used to set the layout (this allows to keep
the statusbar area free from clients). The font is loaded according to
the config, and its information (ascent, descent, height) is used to
correctly position the text in the status bar.

The font and status bar setup is moved to separate functions, and all
setup functions are prefixed with "init" (including setup() -> init()).