Change workspace persistency policy.
Workspaces get attached immediately upon switching to them, they remain
attached until leaving them. This should hopefully simplify a the WSD,
as we no longer need to check whether the selected workspace is attached
or not; furthermore we can be sure that almost every workspace we
interact has got a name (and thus displaying workpace names on the
status bar or a WSD box is greatly simplified).