Make it almost work without Xinerama...
Yes, this sounds backwards, given that it worked *without* Xinerama and
the goal is to make it work *with* Xinerama, too.

But given the recent transition to the Monitor structure, there were
lots of things that stopped working, especially with the static monitor
etc. So the first step was to make at least the one-monitor-no-Xinerama
case work, and apart from some warning messages it seems to works quite
well.

The diff is one big chaos, and I have no clue what's been changed.