Fix WSM crash issue.
Kept crashing when selecting current workspace from WSM when the current
workspace was empty. Problem was that upon leaving the workspace, it was
destroyed, but since we kept coming back (and before it existed), we
were using a non-existing workspace.

... sounds more complicated than it really was, but anyway, if you are
really interested in what was the problem (which I highly doubt), you
may ask me.