Make floating clients work with Xinerama.
The client position did not take into account that the screen offset
could change (and thus used absolute values). The new client position is
calculated as an offset from the current screen's position (and thus is
independent of what screen it is currently displayed on).

To properly handle clients that are (incompletely) moved from one screen
to the other, we search for the biggest intersection of the client with
a screen.