Add session save.
The entire startup process has to be rearranged in order to make the
session restore work (it's a hell of a mess with dependencies between
different parts of the code).

Debug messages now use the correct format specifier for X windows.

Remove input bar (it hasn't been in stwm since... when?).

WARNING: The restarting process hasn't been tested on multiple monitors
yet.