Move focus handling and killing windows to client
There is an ugly hack for the registered buttons, though: Apparently, in
C it is not possible to define an array in a header that is included by
multiple files. As a consequence, currently the `buttons` array is
defined twice (in karuiwm.c and client.c).

The notion of how to handle user configuration remains yet to be
defined.