[client] Fix send_atom function to indicate error correctly
It returned a boolean (as if it was a predicate), but the calling
function interpreted it as an error code (as it actually was an
imperative).

Furthermore, enforce single point of return.

Note that this applies for 'normal' program flow. If we do validity
checks at the beginning of a function, it is OK to leave the function
prematurely (as long as it's done wisely - so no allocated memory and
shit like that, *that* would demand exception handling with goto).