Completely change floating/fullscreen handling
The problem is that we cannot simply raise/lower windows like that now
and then if we want to keep a properly organised stack. But keeping a
properly organised stack is a pain in the ass, honestly. Currently all
floating windows are on the top, and the tiled windows are below (and
it's represented as such in the order in which the clients are stored in
the client list).

Also, fix some of the TODOs added in the last commit, among others:
* Fix function name prefixes in the main source
* Split floating and fullscreen
  - A client is either tiled or floating
  - A client can have multiple states: normal, fullscreen, scratchpad
* Pass buttons to the client when grabbing them (evading the redundant
  declaration of the `buttons` array in both files)
* Split restacking and arranging (`restack` only affects the internally
  used data structure, while `arrange` propagates the changes to the
  server)