Change signed/unsigned int naming policy, add comments
Whenever there is a cast from `int unsigned` to `int`, what we *really*
mean is a cast from and `unsigned` to a `signed` integer, but to mark
every int as `signed` explicitely seemed cumbersome, so the solution was
to write the 'traditional' int explicitely as `signed` and drop the
`int` also on `unsigned` integers, for consistency's sake.

However, the result is a little odd, for two reasons:

* We want to keep the code as explicit as possible. Structures are not
  typedef'd, and types are written explicitely (so no lonely `long` or
  `short` integers, but explicit `int long` and `int short`). Dropping
  the type keyword for such a basic type as integers is thus highly
  braindead.

* People are not used to see `signed` written where one would except
  `int`. After all, the code should be lean and clean, and such things
  only add to the confusion.

Hence, the new policy is to mark signed integers simply as `int`, except
for when there is a cast where all that changes is the signedness, we
explicitely mark it as `int signed`. This is a little less strict
interpretation of 'consistent', but it's acceptable (and the rule is not
even that complicated).