Add client resizing
It currently does not check size hints correctly (or more precisely,
they *are* checked, however the window wiggles upon resizing from the
top and/or left, as the position also changes).

There is also a bug: we do not pose any conditions on the size of the
window. As it reaches a negative value, the window size jumps up to
(1 << sizeof(int unsigned)) - 1, which is not pretty.

Not pretty.