Improve focus to use the right functions/structure members
The LIST_... macro family is a wrapper around the list_... function
family that already does the right casting.

Furthermore, all desktops have got a pointer to the monitor they're
displayed on (or NULL if they're hidden), so there's no need to scan all
the monitors in order to determine whether a desktop is visible or not.