Outsource action handlers to separate file [DOES NOT COMPILE]
The problem is that X event handlers are required for action handlers to
work correctly (in particular for handling the X events during mouse
actions). We could make the definitions non-static, but the whole
"X event handler" thing is flawed by design.

We need to unify X event handlers and general event handlers. In the
long term, the X event handlers are nothing more than a plugin (although
it may end up being a "core plugin").

... so yes, technically, karuiwm should work without X event handlers,
although it doesn't really do much in that case.

Anyway, I'm going to branch this, implement plugin/module handling
properly (with karuibar as a reference), and come back to this branch
once it's handled correctly (and the X event handlers are implemented in
a module).