Start handling X resource lines manually
Since we define keyboard/mouse button assignments as
`karuiwm.keysym.M-S-k: action:arg`, we need to "walk through" the lines,
which does not seem possible with the given functions.

Also, button and key bindings are structures on their own now. This
simplifies their creation and managemtent.

This is a rather intermediary commit that compiles, but has quite a few
ugly parts.